+++ lib/src/helpers/helpers.dart
@@ -8,6 +8,7 @@
 export 'app_lifecycle_observer_mixin.dart';
 export 'clear_focus_navigator_observer.dart';
 export 'helper_functions.dart';
+export 'ticker_helpers.dart';
 export 'widgets/affinity_widgets.dart';
 export 'widgets/gap.dart';
 export 'widgets/hide_keyboard.dart';